Make swift not hang on error.
Before, if a QueueFunctionThread's function raised an exception, then
its thread would log the exception and exit, leaving the rest of the
jobs in the queue and ensuring that the swift client would hang.

Now, the exception is logged and processing continues, so all the
messages get handled eventually and the client exits.

Change-Id: I43d4df212847a2a85732b304de319ea2cce82ddd