Use Emacs-friendly coding line.
To signal the file encoding to Emacs, one can add a line with

  -*- coding: your-favority-encoding -*-

Python also understands such a line, infact it also understands the
line if it says "encoding" instead of "coding".

However, it is only "coding" that has an effect in Emacs. Furthermore,
since "encoding" is not recognized, Emacs will prompt the user with a
scary looking message when the file is opened:

  The local variable list in __init__.py contains values that may not
  be safe.

Using the correct variable name fixes this.

(The file contain only ASCII characters at the moment, so the line
could alternatively be removed completely.)

Change-Id: Ie6b4f41043bf97bb59e3de403e4794c302d81783