Update from global requirements
This syncs up the requirements with global-requirements on stable/juno.

We need this to cap the dependency on futures to avoid a conflict
with capped futures in the version of taskflow in stable/juno, otherwise
setting up devstack with cinder enabled fails.

Also update .gitreview for stable/juno since this is the first patch
on the new branch.

Also sets up the tox func target, which is basically a cherry pick of
commit 8aff0bda9c35ab96146f4f419803e62710ab5618. Note that while we
make it possible to run the functional tests, we exclude them in
Jenkins runs since they are also excluded from stable/kilo due to
instability in the gate.

Depends-On: I5a2d3f7860aeae6bf442e9372286eedbc290d6a8

Change-Id: I507aade31da19be71900b81880cd5568b351663e
Partial-Bug: #1486576