Fix debug and info option parsing
The debug and info options need to be set before a subcommand method is called,
otherwise they are simply ignored. This is kind of irritating - other options
(for example -U, -A, -K) are usable after a positional command.

This patch fixes this, and commands like these are no longer ignoring --debug or --info:

 swift stat --debug
 swift list container --info

Co-Authored-By: Alistair Coles <alistair.coles@hpe.com>

Change-Id: Ib19b05deef7a015881f1eed4a3946025e16bf922