Do not reveal auth token in swiftclient log messages by default
Currently the swiftclient logs sensitive info in headers when logging
HTTP requests. This patch hides sensitive info in headers such as
'X-Auth-Token' in a similar way to swift itself (we add a
'reveal_sensitive_prefix' configuration to the client).

With this patch, tokens are truncated by removing the specified number
of characters, after which '...' is appended to the logged token to
indicate that it has been redacted.

Also include client.parse_header_string() for safe unicode handling
of header data.

Backport based on commits:

  c3f06417049e17a8d45ee5926c5043cb6c8aa9ef
  4d44dcf36086add13d3353915c014f095ab99c6d
  ce569f46517e10f2ce0d27e9ee0a922ad1d84e2f
  46d817828082105a69d4da53fef2f2fbefc54809
  aa0edd00966237163451fc44cda2c593a5215cbe

Co-Authored-By: Tim Burke <tim.burke@gmail.com>
Co-Authored-By: Alistair Coles <alistair.coles@hpe.com>
Co-Authored-By: Li Cheng <shcli@cn.ibm.com>
Co-Authored-By: Zack M. Davis <zdavis@swiftstack.com>

Change-Id: I71fc5aad23bc076b06f75888c3ea507feffc7b48
Closes-bug: #1516692