Correct the example keystone session code imports
This is a follow up to patch Ia3fd947ff619c11ff0ce474897533dcf7b49d9b3.
One of the imports for the keystone session example was wrong. It was
trying to import v3 like:

  from keystoneauth1 import v3

Whereas in my testing v3 actually exists deeper under identity:

  from keystoneauth1.identity import v3

This patch fixes this import in the example documentation.

Change-Id: Ie31144fe65116d20a36b83a96079a3804bf55f29