Allow _retry to continue if there's an SSLError in --insecure mode.
The existing code was throwing an exception instead of continueing to the next
retry. That may be the proper thing to do for a CertificateError, but SSLError
can also be raised on a connection reset.

The existing logic was very problematic when uploading a large file with many
segments. If one segment has a connection reset, the code would stop attempting
to upload that segment, but swiftclient would wait for the rest of the segments
to complete before getting the results of all the futures and refusing to
create the manifest.

Ideally, the exception handling would be further improved to only "raise" a
CertificateError. With this commit, at least there's a workaround for users that
are OK with --insecure mode, and need to upload large files.