Support pdb in tests better
Not really "better" so much as "at all" - the thing we do with the
capture stderr *everywhere* is probably brilliant - but absolutely not
strictly necessary for every MockHttpTest TestCase and comes with the
annoying overhead of trying to get into a debugger causes tests to hang
inexplicably and you can't even do debug prints in tests!?

Now if you add SWIFTCLIENT_DEBUG=1 to your nose -vsx command you can not
only jump into debugger, but if you're "in the know" you could even get
some stderr print debugging going on!

If you're not "in the know" when you try to pdb.set_trace() the tests
will blow-up for you because we monkeypatch pdb when not in
SWIFTCLIENT_DEBUG mode, you're welcome.

Change-Id: I21298bfd39fe386b5ea19e3a6f4408d8a0459c92