Use a valid default for auth_version
The valid set of values for auth_version does not include
values starting with the 'v'.

In this particular function, the auth_version variable is
only used for comparisons with v3. So, the code worked
correctly. However, let's clean this up in order to reduce
review confusion and defuse possible future landmine in case
of code changes.

Change-Id: I671016d7992a1922b786b4eb8876b3fbb2532e15